@@ -1,8 +1,9 @@
 {
 	"name": "botclient",
-	"description": "A discord botclient made using electron, react and discord.js",
+	"productName": "BotClient",
+	"description": "A discord botclient built with Electron, React and discord.js.",
 	"author": "DarkGuy10",
-	"version": "0.3.0",
+	"version": "0.12.1",
 	"private": true,
 	"license": "MIT",
 	"repository": {
@@ -10,21 +11,25 @@
 		"url": "https://github.com/darkGuy10/BotClient.git"
 	},
 	"dependencies": {
-		"@skyra/discord-components-core": "^3.0.1",
-		"@skyra/discord-components-react": "^3.0.1",
-		"@testing-library/jest-dom": "^5.14.1",
-		"@testing-library/react": "^12.0.0",
-		"@testing-library/user-event": "^13.2.1",
-		"discord.js": "^13.5.1",
+		"@darkguy10/discord-markdown": "^2.6.0",
+		"@skyra/discord-components-core": "^3.3.0",
+		"@skyra/discord-components-react": "^3.3.0",
+		"@testing-library/jest-dom": "^5.16.2",
+		"@testing-library/react": "^12.1.4",
+		"@testing-library/user-event": "^13.5.0",
+		"discord-api-types": "^0.37.67",
+		"discord.js": "^14.14.1",
 		"electron-is-dev": "^2.0.0",
-		"electron-store": "^8.0.1",
-		"markdown-to-jsx": "^7.1.5",
-		"pretty-quick": "^3.1.3",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
-		"react-scripts": "5.0.0",
-		"react-twemoji": "^0.4.0",
-		"web-vitals": "^2.1.0"
+		"electron-log": "^5.1.0",
+		"electron-store": "^8.1.0",
+		"electron-updater": "6.1.7",
+		"html-to-react": "^1.4.8",
+		"markdown-to-jsx": "^7.1.7",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
+		"react-scripts": "^5.0.1",
+		"react-twemoji": "^0.5.0",
+		"web-vitals": "^2.1.4"
 	},
 	"main": "public/electron.js",
 	"homepage": "./",
@@ -36,7 +41,8 @@
 		"electron-build": "electron-builder",
 		"build": "yarn react-build && yarn electron-build",
 		"start": "concurrently -k \"yarn react-start\" \"wait-on tcp:127.0.0.1:3000 && electron .\"",
-		"prepare": "husky install"
+		"lint": "prettier -w .",
+		"postinstall": "electron-builder install-app-deps"
 	},
 	"eslintConfig": {
 		"extends": [
@@ -58,11 +64,11 @@
 	},
 	"devDependencies": {
 		"babel-plugin-module-resolver": "^4.1.0",
-		"concurrently": "^7.0.0",
-		"electron": "^16.0.7",
-		"electron-builder": "^22.14.5",
-		"husky": "^7.0.0",
-		"prettier": "^2.5.1",
-		"wait-on": "^6.0.0"
-	}
+		"concurrently": "^8.2.2",
+		"electron": "^28.1.4",
+		"electron-builder": "24.9.1",
+		"prettier": "^3.2.4",
+		"wait-on": "^7.2.0"
+	},
+	"packageManager": "yarn@4.0.2"
 }
