@@ -31,38 +31,42 @@ class MemberNav extends Component {
 		const { hoisted, online, offline } = orderMembers(members)
 		return (
 			<div className={styles.memberNav}>
-				{[...hoisted].map(([role, members], key) => (
-					<div key={key}>
-						<h2 className={styles.header}>
-							<span>
-								{parseTwemojis(role.name)} — {members.length}{' '}
-							</span>
-						</h2>
-						{members.map((member, key) => (
-							<MemberListItem key={key} member={member} />
+				<aside className={styles.membersWrap}>
+					<div className={styles.members}>
+						{[...hoisted].map(([role, members], key) => (
+							<div key={key}>
+								<h2 className={styles.header}>
+									<span>
+										{parseTwemojis(role.name)} — {members.length}{' '}
+									</span>
+								</h2>
+								{members.map((member, key) => (
+									<MemberListItem key={key} member={member} />
+								))}
+							</div>
 						))}
+						{online.length ? (
+							<div>
+								<h2 className={styles.header}>
+									<span>online — {online.length}</span>
+								</h2>
+								{online.map((member, key) => (
+									<MemberListItem key={key} member={member} />
+								))}
+							</div>
+						) : null}
+						{offline.length ? (
+							<div>
+								<h2 className={styles.header}>
+									<span>offline — {offline.length}</span>
+								</h2>
+								{offline.map((member, key) => (
+									<MemberListItem key={key} member={member} />
+								))}
+							</div>
+						) : null}
 					</div>
-				))}
-				{online.length ? (
-					<div>
-						<h2 className={styles.header}>
-							<span>online — {online.length}</span>
-						</h2>
-						{online.map((member, key) => (
-							<MemberListItem key={key} member={member} />
-						))}
-					</div>
-				) : null}
-				{offline.length ? (
-					<div>
-						<h2 className={styles.header}>
-							<span>offline — {offline.length}</span>
-						</h2>
-						{offline.map((member, key) => (
-							<MemberListItem key={key} member={member} />
-						))}
-					</div>
-				) : null}
+				</aside>
 			</div>
 		)
 	}
@@ -73,15 +77,11 @@ class MemberNav extends Component {
  * @param {Array} members
  */
 const orderMembers = members => {
-	const offline = sortAlphabetically(
-		members.filter(member => !member.presence)
-	)
+	const offline = sortAlphabetically(members.filter(member => !member.presence))
 	const online = sortAlphabetically(
 		members.filter(member => member.presence && !member.isHoisted)
 	)
-	const hoisted = members.filter(
-		member => member.presence && member.isHoisted
-	)
+	const hoisted = members.filter(member => member.presence && member.isHoisted)
 
 	const uniqify = roles => {
 		return roles.filter((value, index, self) => {
