@@ -4,14 +4,29 @@ import AlertManager from './AlertManager/AlertManager'
 import BotTag from './BotTag/BotTag'
 import Login from './Login/Login'
 import Layout from './Layout/Layout'
+import UserSection from './UserSection/UserSection'
 import ChannelNav from './ChannelNav/ChannelNav'
+import DMNav from './DMNav/DMNav'
 import GuildNav from './GuildNav/GuildNav'
-import Main from './Main/Main'
+import GuildListItem from './GuildListItem/GuildListItem'
+import Chat from './Chat/Chat'
 import MemberNav from './MemberNav/MemberNav'
 import MemberListItem from './MemberListItem/MemberListItem'
 import MessageElement from './MessageElement/MessageElement'
 import MessageField from './MessageField/MessageField'
 import UploadElement from './UploadElement/UploadElement'
+import ToggleButton from './ToggleButton/ToggleButton'
+import ToggleElement from './ToggleElement/ToggleElement'
+import UserSettings from './UserSettings/UserSettings'
+import ReplyBar from './ReplyBar/ReplyBar'
+import DMListItem from './DMListItem/DMListItem'
+import ChannelListItem from './ChannelListItem/ChannelListItem'
+import TooltipManager from './TooltipManager/TooltipManager'
+import MessageAction from './MessageAction/MessageAction'
+import StatusIndicator from './StatusIndicator/StatusIndicator'
+import ErrorBoundary from './ErrorBoundary/ErrorBoundary'
+import WordFromDevelopers from './WordFromDevelopers/WordFromDevelopers'
+import ContextMenuManager from './ContextMenuManager/ContextMenuManager'
 
 export {
 	App,
@@ -20,12 +35,27 @@ export {
 	BotTag,
 	Login,
 	Layout,
+	UserSection,
+	DMNav,
 	ChannelNav,
 	GuildNav,
-	Main,
+	GuildListItem,
+	Chat,
 	MemberNav,
 	MemberListItem,
 	MessageElement,
 	MessageField,
 	UploadElement,
+	ToggleButton,
+	ToggleElement,
+	UserSettings,
+	ReplyBar,
+	DMListItem,
+	ChannelListItem,
+	TooltipManager,
+	MessageAction,
+	StatusIndicator,
+	ErrorBoundary,
+	WordFromDevelopers,
+	ContextMenuManager,
 }
